Break down test files into smaller ones
The API tests have been split into multiple files to help organize the
project.
