+++ package.json
@@ -12,7 +12,7 @@
     "lint": "eslint .",
     "migrate-test": "dotenv -e test.env -- prisma migrate dev --name test",
     "seeds": "npm run docker:up && npm run compile && node  ./dist/db_seeding/user_seed.js",
-    "test": "npm run docker:up && npm run compile && dotenv -e test.env -- mocha 'dist/test/**/*.js'"
+    "test": "npm run docker:up && npm run compile && dotenv -e test.env -- mocha 'dist/test/index.js'"
   },
   "author": "Pedro Martins Dietrich",
   "license": "ISC",
@@ -21,7 +21,7 @@
     "@prisma/client": "5.19.1",
     "bcryptjs": "2.4.3",
     "dotenv": "16.4.5",
-    "express": "4.19.2",
+    "express": "4.21.0",
     "graphql": "16.9.0",
     "graphql-http": "1.22.1",
     "jsonwebtoken": "9.0.2"
