+++ src/server/server.ts
@@ -6,17 +6,19 @@ import { formatError } from './server-error.js';
 import { resolvers } from '../resolvers.js';
 import { typeDefs } from '../typedefs.js';
 
-export const startServer = async (port: number): Promise<{ server: ApolloServer; url: string }> => {
+export let url: string;
+
+export const startServer = async (port: number): Promise<{ server: ApolloServer }> => {
   const server = new ApolloServer({
     typeDefs,
     resolvers,
     formatError,
   });
 
-  const { url } = await startStandaloneServer(server, {
+  ({ url } = await startStandaloneServer(server, {
     listen: { port: port },
     context: serverContext,
-  });
+  }));
 
-  return { server: server, url: url };
+  return { server: server };
 };
